fixed: tests that try to create duplicate spells do not spam stderr.
