+++ src/battle.c
@@ -61,6 +61,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <util/base36.h>
 #include <util/bsdstring.h>
 #include <util/language.h>
+#include <util/lists.h>
 #include <util/log.h>
 #include <util/parser.h>
 #include <quicklist.h>
